chore: transition to persistent MariaDB with independent SQL schemas
- Migrated users and transactions services from H2 to MariaDB Docker.
- Set spring.sql.init.mode to 'never' to prevent automatic schema execution on startup.
- Refactored schema.sql files with 'IF NOT EXISTS' for idempotent manual execution.
- Removed DevQueryClient and direct-SQL utilities as persistence is now handled by the DB.
- Updated warmup.sql with MariaDB-specific DATE_FORMAT for start-of-month logic.
- Adjusted UserRepository to handle BigInteger ID generation from MariaDB.
- Included mariaDB-docker.config for standardized local container setup.

+++ users/src/main/java/com/budget/users/repository/UserRepository.java
@@ -15,6 +15,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.sql.PreparedStatement;
 import java.util.List;
 
@@ -60,7 +61,7 @@ public class UserRepository {
             ps.setString(5, address.getZipcode());
             return ps;
         }, key);
-        return key.getKeyAs(Integer.class);
+        return key.getKeyAs(BigInteger.class).intValue();
     }
 
     public FeignLoginDTO login(String username) {
